Change development execution environment for typescript + Some folder restructuring + Update packages
