+++ server/package.json
@@ -4,12 +4,12 @@
   "description": "API server for Mikane",
   "scripts": {
     "build": "rimraf ./build && tsc && npm run copy && cd build && npm ci --omit=dev",
-    "esbuild": "rimraf ./build && node ./esbuild.config.mjs && npm run copyesbuild",
+    "esbuild": "rimraf ./build && tsc --noEmit && node ./esbuild.config.mjs && npm run copyesbuild",
     "copyesbuild": "copyfiles \"./public/**/*\" ./build && copyfiles -u 1 src/api.json ./build",
     "copy": "copyfiles package.json package-lock.json \"./public/**/*\" ./build",
     "start": "node build/index.js",
-    "startts": "npx ts-node src/index.ts",
-    "dev": "npx nodemon src/index.ts",
+    "dev": "concurrently --raw \"tsx watch src/index.ts\" \"tsc --noEmit --watch\"",
+    "typecheck": "tsc --noEmit",
     "test": "vitest run --pool=forks --poolOptions.forks.singleFork --coverage",
     "db": "docker-compose -f test_db/docker-compose-test-db.yml -p mikane_test_db up -d"
   },
@@ -19,7 +19,7 @@
     "express": "^4.18.2",
     "express-session": "^1.17.3",
     "helmet": "^7.1.0",
-    "nodemailer": "^6.9.7",
+    "nodemailer": "^6.9.8",
     "pg": "^8.11.3",
     "swagger-ui-express": "^5.0.0"
   },
@@ -27,24 +27,24 @@
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "@types/express-session": "^1.17.10",
-    "@types/node": "^20.10.4",
+    "@types/node": "^20.11.3",
     "@types/nodemailer": "^6.4.14",
     "@types/pg": "^8.10.9",
-    "@types/supertest": "^2.0.16",
+    "@types/supertest": "^6.0.2",
     "@types/swagger-ui-express": "^4.1.6",
-    "@typescript-eslint/eslint-plugin": "^6.14.0",
-    "@typescript-eslint/parser": "^6.14.0",
-    "@vitest/coverage-v8": "^1.0.4",
+    "@typescript-eslint/eslint-plugin": "^6.19.0",
+    "@typescript-eslint/parser": "^6.19.0",
+    "@vitest/coverage-v8": "^1.2.0",
+    "concurrently": "^8.2.2",
     "copyfiles": "^2.4.1",
-    "esbuild": "^0.19.10",
-    "eslint": "^8.55.0",
-    "nodemailer-mock": "^2.0.2",
-    "nodemon": "^3.0.2",
+    "esbuild": "^0.19.11",
+    "eslint": "^8.56.0",
+    "nodemailer-mock": "^2.0.4",
     "rimraf": "^5.0.5",
-    "supertest": "^6.3.3",
-    "ts-node": "^10.9.2",
+    "supertest": "^6.3.4",
+    "tsx": "^4.7.0",
     "typescript": "^5.3.3",
-    "vitest": "^1.0.4"
+    "vitest": "^1.2.0"
   },
   "engines": {
     "node": "20.x.x"
