+++ server/src/api/users.ts
@@ -6,9 +6,9 @@ import * as ec from "../types/errorCodes";
 import { authCheck } from "../middlewares/authCheck";
 import { authenticate, createHash } from "../utils/auth";
 import { generateKey } from "../utils/generateKey";
-import { isUUID } from "../utils/uuidValidator";
-import { isEmail } from "../utils/emailValidator";
-import { isPhoneNumber } from "../utils/phoneValidator";
+import { isUUID } from "../utils/validators/uuidValidator";
+import { isEmail } from "../utils/validators/emailValidator";
+import { isPhoneNumber } from "../utils/validators/phoneValidator";
 import { sendRegisterAccountEmail } from "../email-services/registerAccount";
 import { sendDeleteAccountEmail } from "../email-services/deleteAccount";
 import { Expense, User } from "../types/types";
