+++ server/src/api/validation.ts
@@ -2,10 +2,10 @@ import express from "express";
 import * as ec from "../types/errorCodes";
 import * as db from "../db/dbValidation";
 import { authCheck } from "../middlewares/authCheck";
-import { isEmail } from "../utils/emailValidator";
-import { isUUID } from "../utils/uuidValidator";
+import { isEmail } from "../utils/validators/emailValidator";
+import { isUUID } from "../utils/validators/uuidValidator";
 import { ErrorExt } from "../types/errorExt";
-import { isPhoneNumber } from "../utils/phoneValidator";
+import { isPhoneNumber } from "../utils/validators/phoneValidator";
 const router = express.Router();
 
 /*
