+++ server/src/utils/validators/emailValidator.ts
